requirements.txt
anyio==4.9.0
asgiref==3.9.1
bidict==0.23.1
certifi==2025.7.14
channels==4.2.2
charset-normalizer==3.4.2
click==8.2.1
dj-database-url==3.0.1
dj-rest-auth==7.0.1
Django==5.2.4
django-allauth==65.10.0
django-channels==0.7.0
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
dotenv==0.9.9
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
mysqlclient==2.2.7
oauthlib==3.3.1
paho-mqtt==2.1.0
pillow==11.3.0
PyJWT==2.9.0
PyMySQL==1.1.1
python-dotenv==1.1.1
python-engineio==4.12.2
python-socketio==5.13.0
redis==6.2.0
requests==2.32.4
requests-oauthlib==2.0.0
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
whitenoise==6.9.0
wsproto==1.2.0
